+++ changelog/new_support_allowed_receivers_for_style_collection_compact.md
@@ -1,0 +1,1 @@
+* [#11908](https://github.com/rubocop/rubocop/issues/11908): Support `AllowedReceivers` for `Style/CollectionMethods`. ([@koic][])
